Fix the problem caused by the hardened compiler options.
Added -fno-pic and -fno-stack-protector to the Makefile's CFLAGS,
and removed the description of the bug from README.md.

Since we're in the early stages of the x86 boot process here, we're
not exactly in the part of life where stack canaries make sense.

Having these was breaking the build (at around the point where we
switch into protected mode) ever since the options --enable-default-pie
and --enable-default-ssp got added to gcc's default config parameters
on Arch and Gentoo some time between when I wrote this and now.